Avoid sending duplicate mails for commits
Store commits we saw in the DB, and don't send duplicate mails
for them. This removes the php-src specific hack that I put in
place previously.

Ideally the distinct flag would tell use whether or not a commit
mail should be sent, but it doesn't work the way we need it to.