Remove the ip-to-country fetching.
This script tries to obtain a database from an URL that returns a 404. This
functionality was only really used for re-directing to a mirror, but we long
ago discontinued the mirror system.

Now all it does is set a COUNTRY cookie with "NA," + IP-address in several
headers, and we should remove that too.